requirements.txt
openai
chromadb
python-dotenv
tiktoken
streamlit
numpy
pandas
matplotlib
pypdf